refactor: move frontmatter dir
