+++ packages/preprocess/src/frontmatter/frontmatter.regex.cjs
@@ -1,0 +1,11 @@
+const frontmatterRegex = /---((?:.|\s)+?)---/;
+/**
+ * @param {string} toCheck
+ * @returns {string | false}
+ */
+const containsFrontmatter = (toCheck) =>
+  frontmatterRegex.exec(toCheck)?.[1] ?? false;
+module.exports = {
+  frontmatterRegex,
+  containsFrontmatter,
+};
